feat: move projects out of teams

+++ packages/backend/src/graphql/projectMembership/mutations/projectMembershipDeleteMutationField.ts
@@ -9,22 +9,7 @@ builder.mutationField('projectMembershipDelete', (t) =>
       userId: t.arg.id(),
       projectId: t.arg.id(),
     },
-    authScopes: async (_source, { projectId, userId }) => {
-      const project = await prisma.project.findUnique({
-        where: { id: projectId.toString() },
-        select: { teamId: true },
-      })
-
-      if (!project) {
-        return false
-      }
-
-      const teamMembership = await prisma.teamMembership.findUnique({
-        select: { id: true },
-        where: { userId_teamId: { userId: userId.toString(), teamId: project.teamId } },
-      })
-      return teamMembership ? { isTeamAdminByTeamId: project.teamId } : false
-    },
+    authScopes: (_source, { projectId }) => ({ isProjectMember: projectId.toString() }),
     resolve: async (query, _source, { userId, projectId }) => {
       const projectMembership = await prisma.projectMembership.findUnique({
         where: { userId_projectId: { projectId: projectId.toString(), userId: userId.toString() } },
@@ -32,6 +17,14 @@ builder.mutationField('projectMembershipDelete', (t) =>
       if (!projectMembership) {
         throw new Error('project membership not found')
       }
+
+      const numberOfMembershipsOnTheProject = await prisma.projectMembership.count({
+        where: { projectId: projectId.toString() },
+      })
+      if (numberOfMembershipsOnTheProject === 1) {
+        throw new Error('Membership can not be deleted because user is the last member')
+      }
+
       await prisma.projectMembership.delete({
         where: { userId_projectId: { projectId: projectId.toString(), userId: userId.toString() } },
       })
